Create TestHelper with common test functions
